requirements.txt
# GitHub API
PyGithub>=2.1.1

# Cloud Storage
boto3>=1.26.0  # AWS S3
google-api-python-client>=2.86.0  # Google Drive
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.11.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Utilidades
click>=8.1.3
rich>=13.3.5
cryptography>=41.0.0
pyyaml>=6.0.1
tqdm>=4.65.0
aiofiles>=23.1.0  # Para manejo asíncrono de archivos
python-magic>=0.4.27  # Para detección de tipos MIME

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0 